removing ui file
