+++ applications/auto_setup/frontend/utils.py
@@ -1,0 +1,11 @@
+import dataclasses
+
+from PyQt5.QtWidgets import QCheckBox
+
+
+@dataclasses.dataclass
+class Settings:
+    ssa_cal_checkbox: QCheckBox
+    auto_tune_checkbox: QCheckBox
+    cav_char_checkbox: QCheckBox
+    rf_ramp_checkbox: QCheckBox
