+++ tests/applications/auto_setup/backend/test_setup_cavity.py
