+++ tests/applications/auto_setup/backend/test_setup_linac.py
